完整代码提交

+++ sky-server/src/main/java/com/sky/service/impl/OrdersServiceImpl.java
@@ -2,6 +2,7 @@ package com.sky.service.impl;
 
 import com.alibaba.druid.util.StringUtils;
 import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
@@ -21,6 +22,7 @@ import com.sky.result.Result;
 import com.sky.service.OrderDetailService;
 import com.sky.service.OrdersService;
 import com.sky.service.ShoppingCartService;
+import com.sky.utils.HttpClientUtil;
 import com.sky.utils.WeChatPayUtil;
 import com.sky.vo.OrderPaymentVO;
 import com.sky.vo.OrderStatisticsVO;
@@ -30,6 +32,7 @@ import com.sky.websocket.WebSocketServer;
 import io.swagger.models.auth.In;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -49,6 +52,12 @@ import java.util.stream.Collectors;
 @Service
 public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
     implements OrdersService {
+
+    @Value("${sky.shop.address}")
+    private String shopAddress;
+
+    @Value("${sky.baidu.ak}")
+    private String ak;
 
     @Autowired
     private OrdersMapper ordersMapper;
@@ -76,6 +85,10 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
         if(addressBook == null){
             throw  new AddressBookBusinessException(MessageConstant.ADDRESS_BOOK_IS_NULL);
         }
+
+        checkOutOfRange(addressBook.getCityName()+addressBook.getDistrictName()+addressBook.getDetail());
+
+
         //2.购物车数据为空
         List<ShoppingCart> shoppingCartList =
                 shoppingCartService.list(new LambdaQueryWrapper<ShoppingCart>().eq(ShoppingCart::getUserId,
@@ -83,6 +96,8 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
         if(shoppingCartList == null || shoppingCartList.size() == 0){
             throw new ShoppingCartBusinessException(MessageConstant.SHOPPING_CART_IS_NULL);
         }
+
+
 
 
         //向订单表插入一条数据
@@ -114,8 +129,8 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
         });
         boolean flag = orderDetailService.saveBatch(orderDetails);
 
-        //下单成功，清空用户的购物车数据
-        if(orderRows>0 && flag ){
+        //下单成功(且订单已支付)，清空用户的购物车数据
+        if(orderRows>0 && flag && orders.getStatus() > OrderStatus.PENDING_PAYMENT){
             shoppingCartService.remove(new LambdaQueryWrapper<ShoppingCart>().eq(ShoppingCart::getUserId,
                     BaseContext.getCurrentId()));
         }
@@ -265,6 +280,8 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
     @Override
     public Result repetition(Long id) {
 
+
+
         List<OrderDetail> orderDetailList =
                 orderDetailService.list(new LambdaQueryWrapper<OrderDetail>().eq(OrderDetail::getOrderId, id));;
         ArrayList<ShoppingCart> shoppingCarts = new ArrayList<>();
@@ -314,7 +331,8 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
                         ordersPageQueryDTO.getPhone())
                 .eq(null != ordersPageQueryDTO.getStatus(),Orders::getStatus,ordersPageQueryDTO.getStatus())
                 .between(null != ordersPageQueryDTO.getBeginTime() && null != ordersPageQueryDTO.getEndTime(),
-                        Orders::getOrderTime, ordersPageQueryDTO.getBeginTime(), ordersPageQueryDTO.getEndTime());
+                        Orders::getOrderTime, ordersPageQueryDTO.getBeginTime(), ordersPageQueryDTO.getEndTime())
+                .orderByDesc(Orders::getOrderTime);
 
         Page<Orders> page = new Page<>(ordersPageQueryDTO.getPage(), ordersPageQueryDTO.getPageSize());
         Page<Orders> ordersPage = ordersMapper.selectPage(page, queryWrapper);
@@ -479,6 +497,71 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders>
 
         // 将该订单对应的所有菜品信息拼接在一起
         return String.join("", orderDishList);
+    }
+
+    /**
+     * 检查客户的收货地址是否超出配送范围
+     *
+     * @param address
+     */
+    private void checkOutOfRange(String address) {
+        Map map = new HashMap();
+        map.put("address", shopAddress);
+        map.put("output", "json");
+        map.put("ak", ak);
+
+        //获取店铺的经纬度坐标
+        String shopCoordinate = HttpClientUtil.doGet("https://api.map.baidu.com/geocoding/v3", map);
+
+        JSONObject jsonObject = JSON.parseObject(shopCoordinate);
+        if (!jsonObject.getString("status").equals("0")) {
+            throw new OrderBusinessException("店铺地址解析失败");
+        }
+
+        //数据解析
+        JSONObject location = jsonObject.getJSONObject("result").getJSONObject("location");
+        String lat = location.getString("lat");
+        String lng = location.getString("lng");
+        //店铺经纬度坐标
+        String shopLngLat = lat + "," + lng;
+
+        map.put("address", address);
+        //获取用户收货地址的经纬度坐标
+        String userCoordinate = HttpClientUtil.doGet("https://api.map.baidu.com/geocoding/v3", map);
+
+        jsonObject = JSON.parseObject(userCoordinate);
+        if (!jsonObject.getString("status").equals("0")) {
+            throw new OrderBusinessException("收货地址解析失败");
+        }
+
+        //数据解析
+        location = jsonObject.getJSONObject("result").getJSONObject("location");
+        lat = location.getString("lat");
+        lng = location.getString("lng");
+        //用户收货地址经纬度坐标
+        String userLngLat = lat + "," + lng;
+
+        map.put("origin", shopLngLat);
+        map.put("destination", userLngLat);
+        map.put("steps_info", "0");
+
+        //路线规划
+        String json = HttpClientUtil.doGet("https://api.map.baidu.com/directionlite/v1/driving", map);
+
+        jsonObject = JSON.parseObject(json);
+        if (!jsonObject.getString("status").equals("0")) {
+            throw new OrderBusinessException("配送路线规划失败");
+        }
+
+        //数据解析
+        JSONObject result = jsonObject.getJSONObject("result");
+        JSONArray jsonArray = (JSONArray) result.get("routes");
+        Integer distance = (Integer) ((JSONObject) jsonArray.get(0)).get("distance");
+
+        if (distance > 5000) {
+            //配送距离超过5000米
+            throw new OrderBusinessException("超出配送范围");
+        }
     }
 }
